requirements.txt
discord.py[voice]>=2.3.2
openai>=1.0.0
firebase-admin>=6.5.0
python-dateutil>=2.8.2
pytz>=2023.3
PyNaCl>=1.5.0
SpeechRecognition>=3.10.0
pydub>=0.25.1
